The third day...
Added more easter eggs. Also, changed forgive variables to integers instead of doubles.

+++ Calculator/src/CalculatorSetup.java
@@ -2,12 +2,10 @@ import java.util.Scanner;
 
 public class CalculatorSetup {
 	EasterEggs ee = new EasterEggs();
-
 	double x;
 	double y;
 	double z;
-
-	double forgive = 1;
+	int forgive = 1;
 	String cont = "Y";
 	double rslt;
 	Scanner scanner = new Scanner(System.in);
